bots/skud_bot/requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysql-connector-python==8.0.18
pkg-resources==0.0.0
protobuf==3.10.0
pylint==2.3.1
python-engineio==3.9.3
python-socketio==4.3.1
redis==3.3.11
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.5
websocket-client==0.56.0
wrapt==1.11.2
